requirements.txt
torch~=2.2.0
numpy~=1.26.3
matplotlib~=3.8.0
torchvision~=0.17.0
pillow~=10.2.0
overrides~=7.4.0
click~=8.1.7
scikit-learn~=1.5.0
seaborn
pyyaml~=6.0.1
tqdm~=4.65.0
seaborn~=0.12.2
pandas~=2.1.4
SimpleITK~=2.3.1
einops~=0.7.0
albumentations~=1.4.8
setuptools~=68.2.2
opencv-python